Move db worker actions into separate module
