+++ worker/db-worker-actions/action-all.js
@@ -1,0 +1,5 @@
+'use strict'
+
+module.exports = (db, sql, params) => {
+  return db.prepare(sql).all(...params)
+}
